io_uring: acquire 'mm' for task_work for SQPOLL
If we're unlucky with timing, we could be running task_work after
having dropped the memory context in the sq thread. Since dropping
the context requires a runnable task state, we cannot reliably drop
it as part of our check-for-work loop in io_sq_thread(). Instead,
abstract out the mm acquire for the sq thread into a helper, and call
it from the async task work handler.

Cc: stable@vger.kernel.org # v5.7
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Backported-by: Sergey Temerkhanov <s.temerkhanov@gmail.com>

Original-commit-id: 9d8426a09195e2dcf2aa249de2aaadd792d491c7

Conflicts:
	fs/io_uring.c